Revert "split miner and sharder update into mn-update-node-settings"
