+++ go.mod
@@ -1,8 +1,7 @@
 module github.com/0chain/zwalletcli
 
 require (
-	github.com/0chain/gosdk v1.2.82-0.20210815095507-5fe3bb1f4f4a
-	github.com/lithammer/shortuuid v3.0.0+incompatible // indirect
+	github.com/0chain/gosdk v0.0.0
 	github.com/mattn/go-colorable v0.1.4 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/olekukonko/tablewriter v0.0.5
